Replace BackgroundTask with SwingWorker
Since "move to Java 6" happened a while ago.
